move the inspectClickedLocation function to a new classs called drill down
