+++ src/js/map/map.js
@@ -1,6 +1,5 @@
 
 import L from "leaflet";
-import * as turf from '@turf/turf';
 import "proj4leaflet";
 import {getWFSurl, getWMSurl} from "../helpers/hackneyGeoserver";
 import {
@@ -63,6 +62,7 @@ class Map {
     this.blpuPolygon = null;
     this.geoserver_wfs_url = getWFSurl();
     this.geoserver_wms_url = getWMSurl();
+    this.drilldown = null;
   }
 
   init() {
@@ -208,10 +208,7 @@ class Map {
           this.createMap();
           this.createMapContent();
         }
-        //Capture point clicked from event passed to click handler
-        this.map.on('click', (e) => { 
-         this.inspectClickedLocation(e);
-        });
+       
       })
       .catch(error => {
         console.log(error);
@@ -458,99 +455,99 @@ class Map {
     this.container.insertAdjacentElement("beforeend", element);
   }
 
-  inspectClickedLocation(e){
-    //Create bounding box for the point (L.latLngBounds)
-    var turfClickPoint = turf.point([e.latlng.lng, e.latlng.lat]);
-    //var clickLatLon = L.latLng(e.latlng.lat, e.latlng.lng);
-    //var clickBounds = clickLatLon.toBounds(30);
-    //var clickLatLonBounds = L.latLngBounds(e.latlng.lat, e.latlng.lng);
-    //Empty intersectingFeatures array where we will push the features that intersect the clicked location
-    var intersectingFeatures = [];
-    //Empty globalPopUp variable where we will be adding the popUp content of the intersecting features
-    var globalPopUp = '';
+  // inspectClickedLocation(e){
+  //   //Create bounding box for the point (L.latLngBounds)
+  //   var turfClickPoint = turf.point([e.latlng.lng, e.latlng.lat]);
+  //   //var clickLatLon = L.latLng(e.latlng.lat, e.latlng.lng);
+  //   //var clickBounds = clickLatLon.toBounds(30);
+  //   //var clickLatLonBounds = L.latLngBounds(e.latlng.lat, e.latlng.lng);
+  //   //Empty intersectingFeatures array where we will push the features that intersect the clicked location
+  //   var intersectingFeatures = [];
+  //   //Empty globalPopUp variable where we will be adding the popUp content of the intersecting features
+  //   var globalPopUp = '';
    
-    this.map.eachLayer(layer => {
-      if (layer.feature){
-        // console.log('I am a: ' + layer.feature.geometry.type);  
-        //If the feature is a polygon...
-        if (layer.feature.geometry.type == 'Polygon'){
-          let turfPoly = turf.polygon(layer.feature.geometry.coordinates);
-          //Check if the point is within the polygon. If it is, we push the feature into the intersectingFeatures array and create get the popUp content
-          if (turf.booleanPointInPolygon(turfClickPoint, turfPoly)){
-            intersectingFeatures.push(layer);
-            if (globalPopUp != ''){
-              globalPopUp += '<br/><hr><br/>';
-            }
-            globalPopUp += layer.getPopup().getContent();
-          }
-        }
-        //If the feature is a multipolygon...
-        else if (layer.feature.geometry.type == 'MultiPolygon'){
-          layer.feature.geometry.coordinates.forEach(part => {
-            let turfPoly = turf.polygon(part);
-            if (turf.booleanPointInPolygon(turfClickPoint, turfPoly)){
-              intersectingFeatures.push(layer);
-              if (globalPopUp != ''){
-                globalPopUp += '<br/><hr><br/>';
-              }
-              globalPopUp += layer.getPopup().getContent();
-            }
-          })         
-        }
-        //If the feature is a point or multipoints...
-        else if (layer.feature.geometry.type == 'Point'|| layer.feature.geometry.type == 'MultiPoint'){
-          let turfPoint = turf.point(layer.feature.geometry.coordinates);
-          let turfClickPointBuffer = turf.buffer(turfClickPoint, 0.01, {units: 'kilometers'});
+  //   this.map.eachLayer(layer => {
+  //     if (layer.feature){
+  //       // console.log('I am a: ' + layer.feature.geometry.type);  
+  //       //If the feature is a polygon...
+  //       if (layer.feature.geometry.type == 'Polygon'){
+  //         let turfPoly = turf.polygon(layer.feature.geometry.coordinates);
+  //         //Check if the point is within the polygon. If it is, we push the feature into the intersectingFeatures array and create get the popUp content
+  //         if (turf.booleanPointInPolygon(turfClickPoint, turfPoly)){
+  //           intersectingFeatures.push(layer);
+  //           if (globalPopUp != ''){
+  //             globalPopUp += '<br/><hr><br/>';
+  //           }
+  //           globalPopUp += layer.getPopup().getContent();
+  //         }
+  //       }
+  //       //If the feature is a multipolygon...
+  //       else if (layer.feature.geometry.type == 'MultiPolygon'){
+  //         layer.feature.geometry.coordinates.forEach(part => {
+  //           let turfPoly = turf.polygon(part);
+  //           if (turf.booleanPointInPolygon(turfClickPoint, turfPoly)){
+  //             intersectingFeatures.push(layer);
+  //             if (globalPopUp != ''){
+  //               globalPopUp += '<br/><hr><br/>';
+  //             }
+  //             globalPopUp += layer.getPopup().getContent();
+  //           }
+  //         })         
+  //       }
+  //       //If the feature is a point or multipoints...
+  //       else if (layer.feature.geometry.type == 'Point'|| layer.feature.geometry.type == 'MultiPoint'){
+  //         let turfPoint = turf.point(layer.feature.geometry.coordinates);
+  //         let turfClickPointBuffer = turf.buffer(turfClickPoint, 0.01, {units: 'kilometers'});
           
-          //Check if both buffers intersect. If they do, we push the feature into the intersectingFeatures array and create get the popUp content
-          //if (turf.booleanIntersects(featurePointTurfBuffer, turfClickPointBuffer)){
-          //Check if the feature point is whithin the clicked location buffer. If it is, we push the feature into the intersectingFeatures array and create get the popUp content
-          if (turf.booleanPointInPolygon(turfPoint,turfClickPointBuffer)){
-            intersectingFeatures.push(layer);
-            if (globalPopUp != ''){
-              globalPopUp += '<br/><hr><br/>';
-            }
-            globalPopUp += layer.getPopup().getContent();
-          }       
-        }
-        //If the feature is a line...
-        else if(layer.feature.geometry.type == 'LineString'){
-          //Create a turf line from the feature, create the bounding box and covert the bounding box into a turf polygon. 
-          let turfLine = turf.lineString(layer.feature.geometry.coordinates);
-          //Create a buffer around the clickLatLong point (the clicked location)
-          let clickLatLonBuffer = turf.buffer(turfClickPoint, 0.02, {units: 'kilometers'});
-          //Check if both polygons intersect. If they do, we push the feature into the intersectingFeatures array and create get the popUp content
-          if (turf.lineIntersect(clickLatLonBuffer, turfLine).features.length > 0) {
-            intersectingFeatures.push(layer);
-            if (globalPopUp != ''){
-              globalPopUp += '<br/><hr><br/>';
-            }
-            globalPopUp += layer.getPopup().getContent();
-          }
-        }
-        //If the feature is a multiline...
-        else if (layer.feature.geometry.type == 'MultiLineString'){
-          let clickLatLonBuffer = turf.buffer(turfClickPoint, 0.02, {units: 'kilometers'});
-          layer.feature.geometry.coordinates.forEach(part => {
-            let turfLine = turf.lineString(part);
-            if (turf.lineIntersect(clickLatLonBuffer, turfLine).features.length > 0) {
-              intersectingFeatures.push(layer);
-              if (globalPopUp != ''){
-                globalPopUp += '<br/><hr><br/>';
-              }
-              globalPopUp += layer.getPopup().getContent();
-            }
-          })
-        }       
-      }
-    });
-    if (intersectingFeatures.length > 0){
-      this.map.openPopup(globalPopUp, e.latlng, {
-        offset: L.point(0, -24),
-        maxHeight: 300
-      }); 
-    }
-  }
+  //         //Check if both buffers intersect. If they do, we push the feature into the intersectingFeatures array and create get the popUp content
+  //         //if (turf.booleanIntersects(featurePointTurfBuffer, turfClickPointBuffer)){
+  //         //Check if the feature point is whithin the clicked location buffer. If it is, we push the feature into the intersectingFeatures array and create get the popUp content
+  //         if (turf.booleanPointInPolygon(turfPoint,turfClickPointBuffer)){
+  //           intersectingFeatures.push(layer);
+  //           if (globalPopUp != ''){
+  //             globalPopUp += '<br/><hr><br/>';
+  //           }
+  //           globalPopUp += layer.getPopup().getContent();
+  //         }       
+  //       }
+  //       //If the feature is a line...
+  //       else if(layer.feature.geometry.type == 'LineString'){
+  //         //Create a turf line from the feature, create the bounding box and covert the bounding box into a turf polygon. 
+  //         let turfLine = turf.lineString(layer.feature.geometry.coordinates);
+  //         //Create a buffer around the clickLatLong point (the clicked location)
+  //         let clickLatLonBuffer = turf.buffer(turfClickPoint, 0.02, {units: 'kilometers'});
+  //         //Check if both polygons intersect. If they do, we push the feature into the intersectingFeatures array and create get the popUp content
+  //         if (turf.lineIntersect(clickLatLonBuffer, turfLine).features.length > 0) {
+  //           intersectingFeatures.push(layer);
+  //           if (globalPopUp != ''){
+  //             globalPopUp += '<br/><hr><br/>';
+  //           }
+  //           globalPopUp += layer.getPopup().getContent();
+  //         }
+  //       }
+  //       //If the feature is a multiline...
+  //       else if (layer.feature.geometry.type == 'MultiLineString'){
+  //         let clickLatLonBuffer = turf.buffer(turfClickPoint, 0.02, {units: 'kilometers'});
+  //         layer.feature.geometry.coordinates.forEach(part => {
+  //           let turfLine = turf.lineString(part);
+  //           if (turf.lineIntersect(clickLatLonBuffer, turfLine).features.length > 0) {
+  //             intersectingFeatures.push(layer);
+  //             if (globalPopUp != ''){
+  //               globalPopUp += '<br/><hr><br/>';
+  //             }
+  //             globalPopUp += layer.getPopup().getContent();
+  //           }
+  //         })
+  //       }       
+  //     }
+  //   });
+  //   if (intersectingFeatures.length > 0){
+  //     this.map.openPopup(globalPopUp, e.latlng, {
+  //       offset: L.point(0, -24),
+  //       maxHeight: 300
+  //     }); 
+  //   }
+  // }
 }
 
 export default Map;
